requirements.txt
streamlit-aggrid
extra_streamlit_components
pandas
numpy